epiread.c OK (#410)
* epiread.c OK

* uintptr_t

+++ src/libultra/io/epirawread.c
@@ -1,6 +1,6 @@
 #include "global.h"
 
-s32 __osEPiRawReadIo(OSPiHandle* handle, u32 devAddr, u32* data) {
+s32 __osEPiRawReadIo(OSPiHandle* handle, uintptr_t devAddr, u32* data) {
     s32 status;
     OSPiHandle* curHandle;
